iwlwifi-extra: update list of forced inclusions
Recent Intel bluetooth chipsets are not working, due to missing
firmware (e.g. ibt-0180-0041 for a Meteor Lake based Intel NUC 14).

Updating the list based on the file list from the firmware-iwlwifi
package in debian sid (and bookworm-backports)[0] should fix this
and a few related issues.

Follows commit
 bebdf38 ("iwl 5.19 extra: force some ibt (intel bluetooth) fw inclusion")

[0] https://packages.debian.org/sid/all/firmware-iwlwifi/filelist

Reported-by: Moayad Almalat <m.almalat@proxmox.com>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
--
hope I'm not overlooking a more sophisticated process for compiling
this list (I mixed and matched the output of dpkg -c from debian with
ours)

 fwlist-iwlwifi-extra | 55 ++++++++++++++++++++++++++++++++++++++++++++
 1 file changed, 55 insertions(+)